configs: don't inline generate dynamicconfig when outside a repo
Summary:
When Mercurial is outside a repo and you run a repo-required command,
it will try to construct the localrepo and fail. Unfortunately the dynamicconfig
loading logic treated it like a repo-creation event, and tried to load the
dynamicconfig in memory, expecting the repo to be created.

Since we do an http request during this, if we were offline it would hang
for 30 seconds every single time.  Let's avoid this code path unless creating a
repo.

A separate diff will lower the 30 second timeout.

Reviewed By: kulshrax

Differential Revision: D22710370

fbshipit-source-id: e341e9230d2fdba80059ca086f0f12494a10c5d6