megarepolib: introduce stacked pre-merge deletes
Summary:
To gradually merge one repo into the other, we need to produce multiple slices of the working copy. The sum of these slices has to be equal to
the whole of the original repo's working copy. To create each of these slices all files but the ones in the slice need to be deleted from the working copy.
Before this diff, megarepolib would do this in a single delete commit. This however may be impractical, as it will produce huge commits, which we'll be unable
to process adequately. So this diff essentially introduces gradual deletion for each slice, and calls each slice "deletion stack". This is how it looks (a copy from the docstring):

```
  M1
  . \
  . D11 (https://github.com/facebookexperimental/eden/commit/ac5fca16aef60ac938d85acdde3177df34fdb822)
  .  |
  . D12 (https://github.com/facebookexperimental/eden/commit/4c57c974e3b4323756fe224c70fa22c22fc05a5a)
  .   |
  M2   \
  . \   |
  . D21 (https://github.com/facebookexperimental/eden/commit/1135339320817871e62152539a5153219ffd64b9) |
  .  |  |
  . D22 (https://github.com/facebookexperimental/eden/commit/60419d261b8d915d435e72b92f5b7739fc04f7cb) |
  .   | |
  o    \|
  |     |
  o    PM
  ^     ^
  |      \
main DAG   merged repo's DAG
```
Where:
 - `M1`, `M2` - merge commits, each of which merges only a chunk
   of the merged repo's DAG
 - `PM` is a pre-merge master of the merged repo's DAG
 - `D11 (https://github.com/facebookexperimental/eden/commit/ac5fca16aef60ac938d85acdde3177df34fdb822)`, `D12 (https://github.com/facebookexperimental/eden/commit/4c57c974e3b4323756fe224c70fa22c22fc05a5a)`, `D21 (https://github.com/facebookexperimental/eden/commit/1135339320817871e62152539a5153219ffd64b9)` and `D22 (https://github.com/facebookexperimental/eden/commit/60419d261b8d915d435e72b92f5b7739fc04f7cb)` are commits, which delete
   a chunk of working copy each. Delete commmits are organized
   into delete stacks, so that `D11 (https://github.com/facebookexperimental/eden/commit/ac5fca16aef60ac938d85acdde3177df34fdb822)` and `D12 (https://github.com/facebookexperimental/eden/commit/4c57c974e3b4323756fe224c70fa22c22fc05a5a)` progressively delete
   more and more files.

Reviewed By: StanislavGlebik

Differential Revision: D22778907

fbshipit-source-id: ad0bc31f5901727b6df32f7950053ecdde6f599c