revisionstore: write the lfs blob alongside the pointer
Summary:
The ContentStore code has a strong assumption that all the data fetched ends up
in the hgcache. Unfortunately, this assumption breaks somewhat if an LFS
pointer is in the local store but the blob isn't alonside it. This can happen
for instance when ubundling a bundle that contains a pointer, the pointer will
be written to the local store, but the blob would be fetched in the shared
store.

We can break this assumption a bit in the LFS store code by writing the fetched
blob alongside the pointer, this allows the `get` operation to find both the
pointer and the blob in the same store.

Reviewed By: DurhamG

Differential Revision: D22714708

fbshipit-source-id: 01aedf04d692c787b7cddb0f7a76828ea37dcf29