remotefilelog: ignore inaccessible files on repack
Summary:
Now that packfiles are marked with FILE_DELETE_ON_CLOSE, they can no longer be
opened on Windows, and thus trying to stat them will fail with a permission
denied, failing repack.

This really should only be happening when using EdenFS on Windows, which only a
handful (though growing) number of people are using.

Reviewed By: quark-zju

Differential Revision: D22801408

fbshipit-source-id: f4229e90ce076a65994fb9d193d00c309377323a