changelog: use Rust RevlogIndex for addrevision
Summary:
This makes the Rust code path take care of commit writing.

The feature cannot be enabled yet because the `nodemap` backed by the C index
is no longer aware of new in-memory commits. The next diff migrates nodemap to
be backed by Rust and can turn on this feature altogether.

Reviewed By: DurhamG

Differential Revision: D22657191

fbshipit-source-id: 5f1a60f0b391b06fcd61d10676e2e095f8b7c9d6