remotefilelog: do not allow full prefetch on eden
Summary:
Eden runs without a sparse checkout, so "hg prefetch" can be a very expensive
operation on eden. Let's not allow "hg prefetch" on eden if pats are not
specified - this will force users to specify what exactly they want to
prefetch. Obviously they can still request everything, but this diff will make
it harder to do so.

Reviewed By: DurhamG

Differential Revision: D22946092

fbshipit-source-id: 895505bb90980a74b31ded4a75d102c527801652